+++ Core/Src/main.c
@@ -200,7 +200,7 @@ u8 key_state_last=0;
 float initialAngle = 0.0f;    // 初始航向角
 u8 manual_mode = 0;           // 手动设置角度标志位
 u8 BEPP_ON_flag = 0;          // 蜂鸣器标志位
-u8 BEEP_start_time=0;      // 蜂鸣器开启时间
+u32 BEEP_start_time=0;      // 蜂鸣器开启时间
 /**************enum define****************/
 enum currentPosition
 {
